move schedule and asset functions into their own files
