+++ python_modules/libraries/dagster-airflow/dagster_airflow/__init__.py
@@ -1,8 +1,11 @@
 from airflow.plugins_manager import AirflowPlugin
 from dagster._core.utils import check_dagster_package_version
 
-from .dagster_job_factory import (
+from .dagster_asset_factory import (
     load_assets_from_airflow_dag as load_assets_from_airflow_dag,
+)
+
+from .dagster_job_factory import (
     make_dagster_definitions_from_airflow_dag_bag as make_dagster_definitions_from_airflow_dag_bag,
     make_dagster_definitions_from_airflow_dags_path as make_dagster_definitions_from_airflow_dags_path,
     make_dagster_definitions_from_airflow_example_dags as make_dagster_definitions_from_airflow_example_dags,
